# Conflicts:
#	lib/ui/themes/layouts/components/all.dart
#	lib/ui/themes/layouts/layout.dart

+++ lib/ui/pages/menu/my_photos/photo_page.dart
@@ -78,11 +78,12 @@ class PhotoPage extends StatelessWidget {
                       );
                     },
                   ),
-                  TextAndOrIconActionButtonLight(
-                    translate.delete,
-                    AbiliaIcons.deleteAllClear,
-                    onPressed: () => _deletePhoto(context, sortable),
-                  ),
+                  if (!isInPhotoCalendar)
+                    TextAndOrIconActionButtonLight(
+                      translate.delete,
+                      AbiliaIcons.deleteAllClear,
+                      onPressed: () => _deletePhoto(context, sortable),
+                    ),
                   TextAndOrIconActionButtonLight(
                     translate.close,
                     AbiliaIcons.navigationPrevious,
